Fail GitHub actions builds on new compiler warnings, warn about incorrect log format strings

+++ src/nc_connection.c
@@ -312,7 +312,7 @@ conn_put(struct conn *conn)
 void
 conn_init(void)
 {
-    log_debug(LOG_DEBUG, "conn size %d", sizeof(struct conn));
+    log_debug(LOG_DEBUG, "conn size %d", (int)sizeof(struct conn));
     nfree_connq = 0;
     TAILQ_INIT(&free_connq);
 }
